Add files via upload
QPE circuit ---> ibm_brisbane device mapping for the Schlogl model simulations. Qubits in black correspond to qubits in the QPE circuit that have been mapped onto physical qubits on the IBMQ Brisbane machine. The mapping generated is the same for all system volumes reported in the main text (i.e., V = 1.1, 5.5, and 10.5).